requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
sentence-transformers==2.7.0
rank-bm25==0.2.2
tiktoken==0.7.0
anthropic==0.25.0
numpy==1.26.4
scipy==1.13.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
python-dotenv==1.0.1
